backend/user-service/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
mysql-connector-python==8.0.33
PyJWT==2.8.0
bcrypt==4.0.1
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
